Update for CS2

+++ bootstrap.php
@@ -7,18 +7,17 @@
  * @version     3.0
  * @date        21/10/2012
  */
-$check["php"] = (function_exists('version_compare') && version_compare(phpversion(), '5.3.1', '>='));
-$check["php5.4"] = (function_exists('version_compare') && version_compare(phpversion(), '5.4', '>='));
-$check["mysql"] = extension_loaded('mysql');
+$check["php"] = (function_exists('version_compare') && version_compare(phpversion(), '7.4', '>='));
+$check["mysql"] = extension_loaded('mysqli');
+$check["redis"] = extension_loaded('redis');
+$check["json"] = extension_loaded('json');
 $check["spl"] = extension_loaded('spl');
 $check["sockets"] = extension_loaded("sockets");
-$check["pthreads"] = extension_loaded("pthreads");
 
 define('EBOT_DIRECTORY', __DIR__);
 define('APP_ROOT', __DIR__ . DIRECTORY_SEPARATOR);
 
 require_once APP_ROOT . 'vendor' . DIRECTORY_SEPARATOR . 'autoload.php';
-require_once 'steam-condenser.php';
 //require_once APP_ROOT . 'websocket' . DIRECTORY_SEPARATOR . 'websocket.client.php';
 require_once APP_ROOT . 'websocket' . DIRECTORY_SEPARATOR . 'websocket.hotfix.php';
 
@@ -33,19 +32,13 @@ echo "
 
 echo "PHP Compatibility Test" . PHP_EOL;
 echo "-----------------------------------------------------" . PHP_EOL;
-echo "| PHP 5.3.1 or newer    -> required  -> " . ($check["php"] ? ("[\033[0;32m Yes \033[0m] " . phpversion()) : "[\033[0;31m No \033[0m]") . PHP_EOL;
+echo "| PHP 7.4 or newer      -> required  -> " . ($check["php"] ? ("[\033[0;32m Yes \033[0m] " . phpversion()) : "[\033[0;31m No \033[0m]") . PHP_EOL;
 echo "| Standard PHP Library  -> required  -> " . ($check["spl"] ? "[\033[0;32m Yes \033[0m]" : "[\033[0;31m No \033[0m]") . PHP_EOL;
 echo "| MySQL                 -> required  -> " . ($check["mysql"] ? "[\033[0;32m Yes \033[0m]" : "[\033[0;31m No \033[0m]") . PHP_EOL;
 echo "| Sockets               -> required  -> " . ($check["sockets"] ? "[\033[0;32m Yes \033[0m]" : "[\033[0;31m No \033[0m]") . PHP_EOL;
-echo "| pthreads              -> required  -> " . ($check["pthreads"] ? "[\033[0;32m Yes \033[0m]" : "[\033[0;31m No \033[0m]") . PHP_EOL;
+echo "| Redis                 -> required  -> " . ($check["redis"] ? "[\033[0;32m Yes \033[0m]" : "[\033[0;31m No \033[0m]") . PHP_EOL;
+echo "| JSON                  -> required  -> " . ($check["json"] ? "[\033[0;32m Yes \033[0m]" : "[\033[0;31m No \033[0m]") . PHP_EOL;
 echo "-----------------------------------------------------" . PHP_EOL;
-
-if (!$check["php5.4"]) {
-    echo "| We recommand to use PHP5.4 to get better performance !" . PHP_EOL;
-    echo '-----------------------------------------------------' . PHP_EOL;
-}
-
-unset($check["php5.4"]);
 
 if (in_array(false, $check)) {
     echo "| Your php configuration missed, please make sure that you have all feature !" . PHP_EOL;
@@ -70,7 +63,7 @@ gc_enable();
 function handleShutdown() {
     global $webSocketProcess;
 
-    if (PHP_OS == "Linux")
+    if (PHP_OS == "Linux" && $webSocketProcess)
         proc_terminate($webSocketProcess, 9);
 
     $error = error_get_last();
@@ -85,60 +78,6 @@ register_shutdown_function('handleShutdown');
 
 error_reporting(E_ERROR);
 
-class LoggerArray extends Stackable {
-
-    public function run() {
-        
-    }
-
-}
-
-class LogReceiver extends Thread {
-
-    public $shared_array;
-    public $botIp;
-    public $botPort;
-
-    public function __construct($shared_array, $botIp, $botPort) {
-        $this->shared_array = $shared_array;
-        $this->botIp = $botIp;
-        $this->botPort = $botPort;
-        $this->start();
-    }
-
-    public function run() {
-        $socket = socket_create(AF_INET, SOCK_DGRAM, SOL_UDP);
-        if ($socket) {
-            if (socket_bind($socket, $this->botIp, $this->botPort)) {
-                
-            } else {
-                echo "can't bind " . $this->botIp . ":" . $this->botPort . "\n";
-                return;
-            }
-        } else {
-            echo "can't bind " . $this->botIp . ":" . $this->botPort . "\n";
-            return;
-        }
-
-
-        while (true) {
-            $data = "";
-            $int = @socket_recvfrom($socket, $line, 1500, 0, $from, $port);
-            if ($int) {
-                $ip = $from . ":" . $port;
-                $data = $line;
-            } else {
-                usleep(1000);
-            }
-
-            if ($data) {
-                $this->shared_array[] = $ip . "---" . $data;
-            }
-        }
-    }
-
-}
-
 $config = \eBot\Config\Config::getInstance();
 
 if ($config->getNodeStartupMethod() != "none") {
@@ -151,7 +90,7 @@ if ($config->getNodeStartupMethod() != "none") {
             1 => array("file", APP_ROOT . "logs" . DIRECTORY_SEPARATOR . "websocket.log", "a"),
             2 => array("file", APP_ROOT . "logs" . DIRECTORY_SEPARATOR . "websocket.error", "a")
         );
-	    $webSocketProcess = proc_open($config->getNodeStartupMethod() . ' ' . APP_ROOT . 'websocket_server.js ' . \eBot\Config\Config::getInstance()->getBot_ip() . ' ' . \eBot\Config\Config::getInstance()->getBot_port() . ' ' . (\eBot\Config\Config::getInstance()->isSSLEnabled() ? 'TRUE': 'FALSE') . ' ' . \eBot\Config\Config::getInstance()->getSSLCertificatePath() . ' ' . \eBot\Config\Config::getInstance()->getSSLKeyPath(), $descriptorspec, $pipes);
+	    $webSocketProcess = proc_open($config->getNodeStartupMethod() . ' ' . APP_ROOT . 'websocket_server.mjs ' . \eBot\Config\Config::getInstance()->getBot_ip() . ' ' . \eBot\Config\Config::getInstance()->getBot_port() . ' ' . (\eBot\Config\Config::getInstance()->isSSLEnabled() ? 'TRUE': 'FALSE') . ' ' . \eBot\Config\Config::getInstance()->getSSLCertificatePath() . ' ' . \eBot\Config\Config::getInstance()->getSSLKeyPath(), $descriptorspec, $pipes);
         if (is_resource($webSocketProcess)) {
             fclose($pipes[0]);
             usleep(400000);
@@ -172,8 +111,5 @@ if ($config->getNodeStartupMethod() != "none") {
 }
 
 echo '-----------------------------------------------------' . PHP_EOL;
-
-$loggerData = new LoggerArray();
-$thread = new LogReceiver($loggerData, $config->getBot_ip(), $config->getBot_port());
 
 \eBot\Application\Application::getInstance()->run();
